requirements.txt
alpaca-trade-api==2.1.0
pymongo==4.0.1
python-dotenv==1.0.0 